TST Add testing for z2quaternion
This also fixes a bug in that it was using the wrong dtype, np.float64
rather than np.float32.

+++ src/sdrun/util.py
@@ -234,4 +234,4 @@ def get_num_particles(snapshot: Snapshot) -> int:
 
 
 def z2quaternion(theta: np.ndarray) -> np.ndarray:
-    return rowan.from_euler(theta, 0, 0)
+    return rowan.from_euler(theta, 0, 0).astype(np.float32)
